MCR-3445 make MCRSolrFileIndexHandler configurable
