+++ mycore-solr/src/main/resources/components/solr/config/mycore.properties
@@ -82,6 +82,8 @@ MCR.Solr.SolrInputDocument.Path.Factory=org.mycore.solr.index.file.MCRSolrPathDo
 
 MCR.Solr.IndexHandler.Factory=org.mycore.solr.index.handlers.MCRSolrLazyInputDocumentHandlerFactory
 
+MCR.Solr.IndexHandler.FileIndexHandler=org.mycore.solr.index.handlers.stream.MCRSolrDefaultFileIndexHandler
+
 # the stylesheet used to add fields to the <user/> element on mycore side
 # due to a bug in mcr this must be set (even when the stylesheet named in the
 # property only contains the the stylesheet declaration)
